Задание 7/gui+local_bd/requirements.txt
bcrypt==5.0.0
psycopg2-binary==2.9.9